- jwir3 separated a bunch of auxiliary classes from csColladaConvertor.cpp to csColladaClasses 	  to make it easier to work with. 	- jwir3 attempted to get csgeom/triangulate3d.cpp working, but to no avail. 	- jwir3 changed csColladaConvertor::ConvertLighting() to ConvertMaterials(). 	- jwir3 began implementing ConvertMaterials.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2007/collada@27233 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/triangulate3d.h
@@ -20,6 +20,7 @@
 #include "csgeom/trimesh.h"
 #include "csgeom/vector3.h"
 #include "ivaria/reporter.h"
+#include "csgeom/plane3.h"
 
 #ifndef __CS_GEOM_TRIANGULATE_3D_H_
 #define __CS_GEOM_TRIANGULATE_3D_H_
@@ -41,9 +42,10 @@ namespace Geom
 			private:
 				static bool FindVertexGroups(csContour3& poly, csContour3& reflex, csContour3& convex, csArray<size_t>& ears);
 				static bool Snip(csContour3& polygon, csArray<size_t>& ears, const size_t earPoint, csTriangleMesh& addTo);
-				static bool IsConvex(const csContour3& polygon, const size_t index);
+				static bool IsConvex(const csContour3& polygon, const int index);
 				static bool IsContained(const csVector3& testVertex, const csVector3& a, const csVector3& b, const csVector3& c);
 				static bool IsSameSide(const csVector3& p1, const csVector3& p2, const csVector3& a, const csVector3& b);
+
 		}; /* End class Triangulate3D */
 
 } // namespace Geom
